Finished disassembly of server.js into auth/create-user.js and auth/password.js.
