+++ server/src/server.js
@@ -2,6 +2,8 @@
 
 "use strict";
 
+const Config = require('./config');
+
 const akismetLib = require('akismet');
 const AWS = require('aws-sdk');
 AWS.config.set('region', 'us-east-1');
@@ -43,12 +45,10 @@ const _ = require('underscore');
 
 // Re-import disassembled code to promise existing code will work
 const Log = require('./log');
-const yell = Log.yell;
-const fail = Log.fail;
-const userFail = Log.userFail;
 
 const addInRamMetric = require('./utils/metered').addInRamMetric;
 const MPromise = require('./utils/metered').MPromise;
+const yell = Log.yell;
 
 const pg = require('./db/pg-query');
 const pgQuery = pg.pgQuery;
@@ -58,6 +58,17 @@ const pgQueryP_metered = pg.pgQueryP_metered;
 const pgQueryP_metered_readOnly = pg.pgQueryP_metered_readOnly;
 const pgQueryP_readOnly = pg.pgQueryP_readOnly;
 const pgQueryP_readOnly_wRetryIfEmpty = pg.pgQueryP_readOnly_wRetryIfEmpty;
+
+const CreateUser = require('./auth/create-user');
+const Password = require('./auth/password');
+const doSendVerification = CreateUser.doSendVerification;
+const generateAndRegisterZinvite = CreateUser.generateAndRegisterZinvite;
+const generateToken = Password.generateToken;
+const generateTokenP = Password.generateTokenP;
+
+// TODO: Maybe able to remove
+const generateHashedPassword = require('./auth/password').generateHashedPassword;
+const checkPassword = require('./auth/password').checkPassword;
 
 const cookies = require('./utils/cookies');
 const COOKIES = cookies.COOKIES;
@@ -255,7 +266,7 @@ DD.prototype.s = DA.prototype.s = function(k, v) {
 //   return [];
 // }
 
-const domainOverride = process.env.DOMAIN_OVERRIDE || null;
+const domainOverride = Config.domainOverride;
 
 function haltOnTimeout(req, res, next) {
   if (req.timedout) {
@@ -497,6 +508,9 @@ String.prototype.hashCode = function() {
   return hash;
 };
 
+const fail = Log.fail;
+const userFail = Log.userFail;
+
 function initializePolisHelpers() {
   // // If there are any comments which have no votes by the owner, create a PASS vote by the owner.
   // pgQuery("select * from comments", [], function(err, comments) {
@@ -538,19 +552,6 @@ function initializePolisHelpers() {
   //     });
   // });
 
-  function generateHashedPassword(password, callback) {
-    bcrypt.genSalt(12, function(errSalt, salt) {
-      if (errSalt) {
-        return callback("polis_err_salt");
-      }
-      bcrypt.hash(password, salt, function(errHash, hashedPassword) {
-        if (errHash) {
-          return callback("polis_err_hash");
-        }
-        callback(null, hashedPassword);
-      });
-    });
-  }
 
   const polisTypes = Utils.polisTypes;
   const setCookie = cookies.setCookie;
@@ -2003,28 +2004,8 @@ function initializePolisHelpers() {
       });
     });
   }
-  function getServerNameWithProtocol(req) {
-    let server = "https://pol.is";
 
-    if (domainOverride) {
-      server = req.protocol + "://" + domainOverride;
-    }
-    if (devMode) {
-      // usually localhost:5000
-      server = "http://" + req.headers.host;
-    }
-
-    if (req.headers.host.includes("preprod.pol.is")) {
-      server = "https://preprod.pol.is";
-    }
-    if (req.headers.host.includes("embed.pol.is")) {
-      server = "https://embed.pol.is";
-    }
-    if (req.headers.host.includes("survey.pol.is")) {
-      server = "https://survey.pol.is";
-    }
-    return server;
-  }
+  const getServerNameWithProtocol = Config.getServerNameWithProtocol;
 
   function handle_POST_auth_slack_redirect_uri(req, res) {
     const code = req.p.code;
@@ -2355,166 +2336,6 @@ Feel free to reply to this email if you need help.`;
           winston.log("info", otzinviteArray);
           resolve(otzinviteArray);
         });
-    });
-  }
-
-  function generateTokenP(len, pseudoRandomOk) {
-    return new Promise(function(resolve, reject) {
-      generateToken(len, pseudoRandomOk, function(err, token) {
-        if (err) {
-          reject(err);
-        } else {
-          resolve(token);
-        }
-      });
-    });
-  }
-
-  function generateToken(len, pseudoRandomOk, callback) {
-    // TODO store up a buffer of random bytes sampled at random times to reduce predictability. (or see if crypto module does this for us)
-    // TODO if you want more readable tokens, see ReadableIds
-    let gen;
-    if (pseudoRandomOk) {
-      gen = crypto.pseudoRandomBytes;
-    } else {
-      gen = crypto.randomBytes;
-    }
-    gen(len, function(err, buf) {
-      if (err) {
-        return callback(err);
-      }
-
-      let prettyToken = buf.toString('base64')
-        .replace(/\//g, 'A').replace(/\+/g, 'B') // replace url-unsafe tokens (ends up not being a proper encoding since it maps onto A and B. Don't want to use any punctuation.)
-        .replace(/l/g, 'C') // looks like '1'
-        .replace(/L/g, 'D') // looks like '1'
-        .replace(/o/g, 'E') // looks like 0
-        .replace(/O/g, 'F') // looks lke 0
-        .replace(/1/g, 'G') // looks like 'l'
-        .replace(/0/g, 'H') // looks like 'O'
-        .replace(/I/g, 'J') // looks like 'l'
-        .replace(/g/g, 'K') // looks like 'g'
-        .replace(/G/g, 'M') // looks like 'g'
-        .replace(/q/g, 'N') // looks like 'q'
-        .replace(/Q/g, 'R') // looks like 'q'
-      ;
-      // replace first character with a number between 2 and 9 (avoiding 0 and 1 since they look like l and O)
-      prettyToken = _.random(2, 9) + prettyToken.slice(1);
-      prettyToken = prettyToken.toLowerCase();
-      prettyToken = prettyToken.slice(0, len); // in case it's too long
-
-      callback(0, prettyToken);
-    });
-  }
-
-  // function generateApiKeyForUser(uid, optionalPrefix) {
-  //   let parts = ["pkey"];
-  //   let len = 32;
-  //   if (!_.isUndefined(optionalPrefix)) {
-  //     parts.push(optionalPrefix);
-  //   }
-  //   len -= parts[0].length;
-  //   len -= (parts.length - 1); // the underscores
-  //   parts.forEach(function(part) {
-  //     len -= part.length;
-  //   });
-  //   return generateTokenP(len, false).then(function(token) {
-  //     parts.push(token);
-  //     let apikey = parts.join("_");
-  //     return apikey;
-  //   });
-  // }
-
-  // function addApiKeyForUser(uid, optionalPrefix) {
-  //   return generateApiKeyForUser(uid, optionalPrefix).then(function(apikey) {
-  //     return pgQueryP("insert into apikeysndvweifu (uid, apikey)  VALUES ($1, $2);", [uid, apikey]);
-  //   });
-  // }
-  // function getApiKeysTruncated(uid) {
-  //   return pgQueryP_readOnly("select * from apikeysndvweifu WHERE uid = ($1);", [uid]).then(function(rows) {
-  //     if (!rows || !rows.length) {
-  //       return [];
-  //     }
-  //     return rows.map(function(row) {
-  //       return {
-  //         apikeyTruncated: row.apikey.slice(0, 10) + "...",
-  //         created: row.created,
-  //       };
-  //     });
-  //   });
-  // }
-
-  // function createApiKey(uid) {
-  //   return generateTokenP(17, false).then(function(token) {
-  //     let apikey = "pkey_" + token;
-  //     return pgQueryP("insert into apikeysndvweifu (uid, apikey) values ($1, $2) returning *;", [uid, apikey]).then(function(row) {
-  //       return {
-  //         apikey: apikey,
-  //         created: row.created,
-  //       };
-  //     });
-  //   });
-  // }
-
-  // function deleteApiKey(uid, apikeyTruncated) {
-  //   // strip trailing "..."
-  //   apikeyTruncated = apikeyTruncated.slice(0, apikeyTruncated.indexOf("."));
-  //   // basic sanitizing - replace unexpected characters with x's.
-  //   apikeyTruncated = apikeyTruncated.replace(/[^a-zA-Z0-9_]/g, 'x');
-  //   return pgQueryP("delete from apikeysndvweifu where uid = ($1) and apikey ~ '^" + apikeyTruncated + "';", [uid]);
-  // }
-
-  // function addApiKeyForUsersBulk(uids, optionalPrefix) {
-  //     let promises = uids.map(function(uid) {
-  //         return generateApiKeyForUser(uid, optionalPrefix);
-  //     });
-  //     return Promise.all(promises).then(function(apikeys) {
-  //         let query = "insert into apikeysndvweifu (uid, apikey)  VALUES ";
-  //         let pairs = [];
-  //         for (var i = 0; i < uids.length; i++) {
-  //             let uid = uids[i];
-  //             let apikey = apikeys[i];
-  //             pairs.push("(" + uid + ', \'' + apikey + '\')');
-  //         }
-  //         query += pairs.join(',');
-  //         query += 'returning uid;';
-  //         return pgQueryP(query, []);
-  //     });
-  // }
-
-  // let uidsX = [];
-  // for (var i = 200200; i < 300000; i++) {
-  //     uidsX.push(i);
-  // }
-  // addApiKeyForUsersBulk(uidsX, "test23").then(function(uids) {
-  //     console.log("hihihihi", uids.length);
-  //     setTimeout(function() { process.exit();}, 3000);
-  // });
-
-  // // let time1 = Date.now();
-  // createDummyUsersBatch(3 * 1000).then(function(uids) {
-  //         // let time2 = Date.now();
-  //         // let dt = time2 - time1;
-  //         // console.log("time foo" , dt);
-  //         // console.dir(uids);
-  //         uids.forEach(function(uid) {
-  //             console.log("hihihihi", uid);
-  //         });
-  //         process.exit(0);
-
-  // }).catch(function(err) {
-  //     console.error("errorfooooo");
-  //     console.error(err);
-  // });
-  function generateAndRegisterZinvite(zid, generateShort) {
-    let len = 10;
-    if (generateShort) {
-      len = 6;
-    }
-    return generateTokenP(len, false).then(function(zinvite) {
-      return pgQueryP('INSERT INTO zinvites (zid, zinvite, created) VALUES ($1, $2, default);', [zid, zinvite]).then(function(rows) {
-        return zinvite;
-      });
     });
   }
 
@@ -3248,22 +3069,6 @@ Thank you for using Polis`;
       body);
   }
 
-  function sendVerificaionEmail(req, email, einvite) {
-    let serverName = getServerNameWithProtocol(req);
-    let body =
-`Welcome to pol.is!
-
-Click this link to verify your email address:
-
-${serverName}/api/v3/verify?e=${einvite}`;
-
-    return sendTextEmail(
-      POLIS_FROM_ADDRESS,
-      email,
-      "Polis verification",
-      body);
-  }
-
   function isEmailVerified(email) {
     return pg.pgQueryP("select * from email_validations where email = ($1);", [email]).then(function(rows) {
       return rows.length > 0;
@@ -3538,25 +3343,6 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-  function checkPassword(uid, password) {
-    return pgQueryP_readOnly_wRetryIfEmpty("select pwhash from jianiuevyew where uid = ($1);", [uid]).then(function(rows) {
-      if (!rows || !rows.length) {
-        return null;
-      } else if (!rows[0].pwhash) {
-        return void 0;
-      }
-      let hashedPassword = rows[0].pwhash;
-      return new Promise(function(resolve, reject) {
-        bcrypt.compare(password, hashedPassword, function(errCompare, result) {
-          if (errCompare) {
-            reject(errCompare);
-          } else {
-            resolve(result ? "ok" : 0);
-          }
-        });
-      });
-    });
-  }
 
   const addLtiUserIfNeeded = User.addLtiUserIfNeeded;
   const addLtiContextMembership = User.addLtiContextMembership;
@@ -5005,150 +4791,12 @@ Email verified! You can close this tab or hit the back button.
           fail(res, 500, "polis_err_reg_fb_user_misc", err);
         });
     });
+
   } // end do_handle_POST_auth_facebook
   function handle_POST_auth_new(req, res) {
-    let hname = req.p.hname;
-    let password = req.p.password;
-    let password2 = req.p.password2; // for verification
-    let email = req.p.email;
-    let oinvite = req.p.oinvite;
-    let zinvite = req.p.zinvite;
-    let referrer = req.cookies[COOKIES.REFERRER];
-    let organization = req.p.organization;
-    let gatekeeperTosPrivacy = req.p.gatekeeperTosPrivacy;
-    let lti_user_id = req.p.lti_user_id;
-    let lti_user_image = req.p.lti_user_image;
-    let lti_context_id = req.p.lti_context_id;
-    let tool_consumer_instance_guid = req.p.tool_consumer_instance_guid;
-    let afterJoinRedirectUrl = req.p.afterJoinRedirectUrl;
-
-    let site_id = void 0;
-    if (req.p.encodedParams) {
-      let decodedParams = decodeParams(req.p.encodedParams);
-      if (decodedParams.site_id) {
-        // NOTE: we could have just allowed site_id to be passed as a normal param, but then we'd need to think about securing that with some other token sooner.
-        // I think we can get by with this obscure scheme for a bit.
-        // TODO_SECURITY add the extra token associated with the site_id owner.
-        site_id = decodedParams.site_id;
-      }
-    }
-
-    if (password2 && (password !== password2)) {
-      fail(res, 400, "Passwords do not match.");
-      return;
-    }
-    if (!gatekeeperTosPrivacy) {
-      fail(res, 400, "polis_err_reg_need_tos");
-      return;
-    }
-    if (!email) {
-      fail(res, 400, "polis_err_reg_need_email");
-      return;
-    }
-    if (!hname) {
-      fail(res, 400, "polis_err_reg_need_name");
-      return;
-    }
-    if (!password) {
-      fail(res, 400, "polis_err_reg_password");
-      return;
-    }
-    if (password.length < 6) {
-      fail(res, 400, "polis_err_reg_password_too_short");
-      return;
-    }
-    if (!_.contains(email, "@") || email.length < 3) {
-      fail(res, 400, "polis_err_reg_bad_email");
-      return;
-    }
-
-    pgQueryP("SELECT * FROM users WHERE email = ($1)", [email]).then(function(rows) {
-
-      if (rows.length > 0) {
-        fail(res, 403, "polis_err_reg_user_with_that_email_exists");
-        return;
-      }
-
-      generateHashedPassword(password, function(err, hashedPassword) {
-        if (err) {
-          fail(res, 500, "polis_err_generating_hash", err);
-          return;
-        }
-        let query = "insert into users " +
-          "(email, hname, zinvite, oinvite, is_owner" + (site_id ? ", site_id" : "") + ") VALUES " + // TODO use sql query builder
-          "($1, $2, $3, $4, $5" + (site_id ? ", $6" : "") + ") " + // TODO use sql query builder
-          "returning uid;";
-        let vals =
-          [email, hname, zinvite || null, oinvite || null, true];
-        if (site_id) {
-          vals.push(site_id); // TODO use sql query builder
-        }
-
-        pgQuery(query, vals, function(err, result) {
-          if (err) {
-            winston.log("info", err);
-            fail(res, 500, "polis_err_reg_failed_to_add_user_record", err);
-            return;
-          }
-          let uid = result && result.rows && result.rows[0] && result.rows[0].uid;
-
-          pgQuery("insert into jianiuevyew (uid, pwhash) values ($1, $2);", [uid, hashedPassword], function(err, results) {
-            if (err) {
-              winston.log("info", err);
-              fail(res, 500, "polis_err_reg_failed_to_add_user_record", err);
-              return;
-            }
-            startSession(uid, function(err, token) {
-              if (err) {
-                fail(res, 500, "polis_err_reg_failed_to_start_session", err);
-                return;
-              }
-              addCookies(req, res, token, uid).then(function() {
-
-                let ltiUserPromise = lti_user_id ?
-                  addLtiUserIfNeeded(uid, lti_user_id, tool_consumer_instance_guid, lti_user_image) :
-                  Promise.resolve();
-                let ltiContextMembershipPromise = lti_context_id ?
-                  addLtiContextMembership(uid, lti_context_id, tool_consumer_instance_guid) :
-                  Promise.resolve();
-                Promise.all([ltiUserPromise, ltiContextMembershipPromise]).then(function() {
-                  if (lti_user_id) {
-                    if (afterJoinRedirectUrl) {
-                      res.redirect(afterJoinRedirectUrl);
-                    } else {
-                      User.renderLtiLinkageSuccessPage(req, res, {
-                        // may include token here too
-                        context_id: lti_context_id,
-                        uid: uid,
-                        hname: hname,
-                        email: email,
-                      });
-                    }
-                  } else {
-                    res.json({
-                      uid: uid,
-                      hname: hname,
-                      email: email,
-                      // token: token
-                    });
-                  }
-                }).catch(function(err) {
-                  fail(res, 500, "polis_err_creating_user_associating_with_lti_user", err);
-                });
-              }, function(err) {
-                fail(res, 500, "polis_err_adding_cookies", err);
-              }).catch(function(err) {
-                fail(res, 500, "polis_err_adding_user", err);
-              });
-            }); // end startSession
-          }); // end insert pwhash
-        }); // end insert user
-      }); // end generateHashedPassword
-
-    }, function(err) {
-      fail(res, 500, "polis_err_reg_checking_existing_users", err);
-    });
+    CreateUser.createUser(req, res);
   } // end /api/v3/auth/new
+
   function handle_POST_tutorial(req, res) {
     let uid = req.p.uid;
     let step = req.p.step;
@@ -8623,19 +8271,6 @@ Email verified! You can close this tab or hit the back button.
       }
     });
   }
-  function decodeParams(encodedStringifiedJson) {
-    if (!encodedStringifiedJson.match(/^\/?ep1_/)) {
-      throw new Error("wrong encoded params prefix");
-    }
-    if (encodedStringifiedJson[0] === "/") {
-      encodedStringifiedJson = encodedStringifiedJson.slice(5);
-    } else {
-      encodedStringifiedJson = encodedStringifiedJson.slice(4);
-    }
-    let stringifiedJson = hexToStr(encodedStringifiedJson);
-    let o = JSON.parse(stringifiedJson);
-    return o;
-  }
 
   function encodeParams(o) {
     let stringifiedJson = JSON.stringify(o);
@@ -10683,13 +10318,6 @@ Thanks for using Polis!
     });
   }
 
-  function doSendVerification(req, email) {
-    return generateTokenP(30, false).then(function(einvite) {
-      return pgQueryP("insert into einvites (email, einvite) values ($1, $2);", [email, einvite]).then(function(rows) {
-        return sendVerificaionEmail(req, email, einvite);
-      });
-    });
-  }
 /*
 
 CREATE TABLE slack_users (
